start new state estimation implementation
